Create a deprecation path caused by upcoming Symfony 6 changes.
The (services) container should not be used, making it necessary
for any controller to inject the services by their own. This makes
the Page Bundle AbstractController redundant.

Also included that the Page Bundle PageController will be final
in the future as the only use for extending it, was for the handy
functionalities that are now a bit more redundant because of the
above.

Added the PageControllerTrait to still be able to use these
"handy functionalities".